pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drakkar"
version = "1.6.6"
description = "Metagenomics pipeline optimized for Mjolnir."
readme = "README.md"
requires-python = ">=3.6"
authors = [
  {name = "Antton Alberdi", email = "antton.alberdi@sund.ku.dk"},
]
dependencies = [
  "numpy",
  "pandas",
  "PyYAML",
  "plotly",
  "biopython",
  "rich",
]

[project.optional-dependencies]
docs = [
  "sphinx>=8,<9",
  "sphinx-rtd-theme>=3,<4",
]
release = [
  "build>=1.2",
  "twine>=5",
]

[project.scripts]
drakkar = "drakkar.cli:main"

[tool.setuptools.packages.find]
include = ["drakkar*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
drakkar = [
  "profile/*",
  "profile/local/*",
  "profile/slurm/*",
  "workflow/*",
  "workflow/envs/*",
  "workflow/rules/*",
  "workflow/scripts/*",
]
